+++ Sources/Views/ALKContextTitleView.swift
@@ -91,18 +91,17 @@ open class ALKContextTitleView: UIView, ALKContextTitleViewType {
         contextImageView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 10).isActive = true
         contextImageView.topAnchor.constraint(equalTo: view.topAnchor, constant: 20).isActive = true
         contextImageView.bottomAnchor.constraint(equalTo: view.bottomAnchor, constant: -10).isActive = true
-        contextImageView.widthAnchor.constraint(equalTo: contextImageView.heightAnchor).isActive = true
         contextImageView.widthAnchor.constraint(lessThanOrEqualTo: view.widthAnchor, multiplier: 0.2).isActive = true
 
         titleLabel.leadingAnchor.constraint(equalTo: contextImageView.trailingAnchor, constant: 10).isActive = true
         titleLabel.topAnchor.constraint(equalTo: contextImageView.topAnchor).isActive = true
         titleLabel.widthAnchor.constraint(equalTo: view.widthAnchor, multiplier: 0.4).isActive = true
-        titleLabel.heightAnchor.constraint(lessThanOrEqualTo: view.widthAnchor, multiplier: 0.3).isActive = true
+        titleLabel.heightAnchor.constraint(lessThanOrEqualTo: view.heightAnchor, multiplier: 0.3).isActive = true
 
         subtitleLabel.leadingAnchor.constraint(equalTo: contextImageView.trailingAnchor, constant: 10).isActive = true
         subtitleLabel.bottomAnchor.constraint(equalTo: contextImageView.bottomAnchor).isActive = true
         subtitleLabel.widthAnchor.constraint(equalTo: view.widthAnchor, multiplier: 0.4).isActive = true
-        subtitleLabel.heightAnchor.constraint(lessThanOrEqualTo: view.widthAnchor, multiplier: 0.3).isActive = true
+        subtitleLabel.topAnchor.constraint(greaterThanOrEqualTo: titleLabel.bottomAnchor, constant: 10).isActive = true
 
         topRightInfoLabel.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: 10).isActive = true
         topRightInfoLabel.topAnchor.constraint(equalTo: contextImageView.topAnchor).isActive = true