+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 - [AL-3391] Add support for autosuggestions.
 - Added setting to hide contact-share option in chatbar.
 - Added setting to disable swipe action in chat list.
+- [AL-3396] Added support for context-based group_of_two conversation.
 
 ### Fixes
 - [AL-3533] Fixed an issue where scrolling in group detail screen would cause inconsistency in profile images of participants.